Here is where the magic happens
All we need to do now is forward our logging messages to Crashlytics.

First we import the Crashlytics class, so we can use the log function.
Then, all we need to do is just forward messages to Crashlytics when in
production. While in development, we'll just keep the default logging
behavior.