Now, just to replace the default logger with our finished one.
First, we need to tell Fabric we want to use Crashlytics. This is the
same process we would follow when using Crashlytics with Fabric on a
normal Android app, as seen [here](https://docs.fabric.io/android/fabric/integration.html#fabric-with-kits).

After that, we just tell FLog that we want to use our new logger for the
LoggingDelegate. It's preeeeettty simple and straightforward.